refactor: move controller attributes to switch only
