+++ custom_components/opensprinkler/switch.py
@@ -5,6 +5,7 @@ from homeassistant.core import HomeAssistant
 from homeassistant.components.switch import SwitchEntity
 from homeassistant.helpers import entity_platform
 from homeassistant.util import slugify
+from homeassistant.util.dt import utc_from_timestamp
 
 from . import (
     OpenSprinklerControllerEntity,
@@ -83,6 +84,86 @@ class ControllerOperationSwitch(
             return "mdi:barley"
 
         return "mdi:barley-off"
+
+    @property
+    def device_state_attributes(self):
+        controller = self._controller
+        attributes = {"opensprinkler_type": "controller"}
+        for attr in [
+            "firmware_version",
+            "hardware_version",
+            "hardware_type",
+            "last_run_station",
+            "last_run_program",
+            "last_run_duration",
+            "rain_sensor_active",
+            "sensor_1_enabled",
+            "sensor_2_enabled",
+            "rain_sensor_enabled",
+            "flow_sensor_enabled",
+            "soil_sensor_enabled",
+            "last_weather_call",
+            "last_successfull_weather_call",
+            "last_weather_call_error",
+            "last_weather_call_error_name",
+            "last_reboot_time",
+            "last_reboot_cause",
+            "last_reboot_cause_name",
+        ]:
+            try:
+                attributes[attr] = getattr(controller, attr)
+            except:
+                pass
+
+        for attr in [
+            "last_weather_call",
+            "last_successfull_weather_call",
+            "last_reboot_time",
+        ]:
+            iso_attr = attr + "_iso"
+            timestamp = getattr(controller, attr)
+            if not timestamp:
+                attributes[attr] = None
+                attributes[iso_attr] = None
+            else:
+                attributes[iso_attr] = utc_from_timestamp(timestamp).isoformat()
+
+        # station counts
+        attributes["station_total_count"] = len(controller.stations)
+        attributes["station_enabled_count"] = len(
+            dict(filter(lambda e: e[1].enabled == True, controller.stations.items()))
+        )
+        attributes["station_is_running_count"] = len(
+            dict(filter(lambda e: e[1].is_running == True, controller.stations.items()))
+        )
+
+        for status in [
+            "manual",
+            "once_program",
+            "master_engaged",
+            "idle",
+            "program",
+            "waiting",
+        ]:
+            key = f"station_{status}_count"
+            attributes[key] = len(
+                dict(
+                    filter(
+                        lambda e: e[1].status == status, controller.stations.items(),
+                    )
+                )
+            )
+
+        # program counts
+        attributes["program_total_count"] = len(controller.programs)
+        attributes["program_enabled_count"] = len(
+            dict(filter(lambda e: e[1].enabled == True, controller.programs.items()))
+        )
+        attributes["program_is_running_count"] = len(
+            dict(filter(lambda e: e[1].is_running == True, controller.programs.items()))
+        )
+
+        return attributes
 
     def _get_state(self) -> str:
         """Retrieve latest state."""
